refactor: button 관련 class 리팩토링
